add timestamp

+++ src/pages/index.js
@@ -20,10 +20,13 @@ class BlogList {
 
     async onDeleteButtonClick(e) {
         if (e.target.className == 'delete') {
-            const blogId = e.target.dataset.id;
-            // console.log(e.target.dataset)
-            await deleteBlog(blogId);
-            this.render();
+            var r = confirm("Are you sure to delete this blog?");
+            if (r == true) {
+                const blogId = e.target.dataset.id;
+                // console.log(e.target.dataset)
+                await deleteBlog(blogId);
+                this.render();
+            }
         }
     }
 
@@ -40,7 +43,7 @@ class BlogList {
     async render() {
         const blogs = await getBlogs();
         let lis = '';
-        blogs.forEach((blog) => lis += `<li data-id=` + sanitizeHTML(blog.id) + `><span class="name">` + sanitizeHTML(blog.title) + `</span><span class="delete" data-id=` + sanitizeHTML(blog.id) + `>delete</span></li>`);
+        blogs.forEach((blog) => lis += `<li data-id=` + sanitizeHTML(blog.id) + `><span class="name">` + sanitizeHTML(blog.title) + `</span><span class="delete" data-id=` + sanitizeHTML(blog.id) + `>delete</span><span class="date">` + blog.createdAt + `</span></li>`);
         this.list.innerHTML = lis;
     }
 }
